+++ lib/services/group_sync_service.dart
@@ -1,6 +1,7 @@
 import 'dart:async';
 
 import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:cloud_functions/cloud_functions.dart';
 import 'package:flutter/foundation.dart';
 
 import '../models/vocabulary_group.dart';
@@ -16,6 +17,10 @@ class GroupSyncService {
   }
 
   static final GroupSyncService instance = GroupSyncService._();
+
+  final FirebaseFunctions _functions = FirebaseFunctions.instanceFor(
+    region: 'europe-west1',
+  );
 
   final ValueNotifier<List<VocabularyGroup>> _groupsNotifier = ValueNotifier(
     const [],
@@ -125,6 +130,12 @@ class GroupSyncService {
       'deleted': true,
       'deletedAt': FieldValue.serverTimestamp(),
       'updatedAt': FieldValue.serverTimestamp(),
+    });
+  }
+
+  Future<void> hardDeleteGroup(String groupId) async {
+    await _functions.httpsCallable('hardDeleteGroup').call({
+      'groupId': groupId,
     });
   }